ci: fix workspace compile, clippy, and test failures after merge
- Fix family_wallet withdrawal tier logic to use LargeWithdrawal config and
  skip multisig for regular-tier withdrawals
- Initialize rollover spending trackers via get_spending_tracker and align
  member-cleanup tests with current client APIs
- Update savings_goals create_goal call sites for the locked parameter
- Resolve clippy/rustfmt issues across family_wallet, savings_goals,
  reporting, remittance_split, bill_payments, and insurance
- Fix signer_rotation threshold test and reporting trend benchmark counts

+++ reporting/src/tests.rs
@@ -3361,10 +3361,7 @@ fn test_top_n_reports_tie_break_is_deterministic_bills() {
     // Deterministic across repeated calls.
     assert_eq!(r1.items.len(), r2.items.len());
     for i in 0..r1.items.len() {
-        assert_eq!(
-            r1.items.get(i as u32).unwrap().id,
-            r2.items.get(i as u32).unwrap().id
-        );
+        assert_eq!(r1.items.get(i).unwrap().id, r2.items.get(i).unwrap().id);
     }
 
     // All amounts equal => order by id ascending => [1,2,3,4,5] capped to MAX.
@@ -3410,10 +3407,7 @@ fn test_top_n_reports_tie_break_is_deterministic_savings() {
 
     assert_eq!(r1.items.len(), r2.items.len());
     for i in 0..r1.items.len() {
-        assert_eq!(
-            r1.items.get(i as u32).unwrap().id,
-            r2.items.get(i as u32).unwrap().id
-        );
+        assert_eq!(r1.items.get(i).unwrap().id, r2.items.get(i).unwrap().id);
     }
 
     let expected_ids = [1u32, 2, 3, 4, 5];
